fix mfkey size

+++ lib/subghz/protocols/keeloq_common.c
@@ -90,7 +90,7 @@ inline uint64_t
  */
 
 inline uint64_t 
-    subghz_protocol_keeloq_common_faac_learning(const uint32_t seed, const uint32_t key) {
+    subghz_protocol_keeloq_common_faac_learning(const uint32_t seed, const uint64_t key) {
     uint16_t hs = seed >> 16;
     const uint16_t ending = 0x544D;
     uint32_t lsb = (uint32_t)hs << 16 | ending;
