+++ lib/subghz/protocols/keeloq_common.h
@@ -73,4 +73,4 @@ uint64_t subghz_protocol_keeloq_common_magic_xor_type1_learning(uint32_t data, u
  * @param key - mfkey (64bit)
  * @return man_learning for this fix number (64bit)
  */
-uint64_t subghz_protocol_keeloq_common_faac_learning(const uint32_t seed, const uint32_t key);
+uint64_t subghz_protocol_keeloq_common_faac_learning(const uint32_t seed, const uint64_t key);
